% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.rsv_plot_daily}
\alias{.rsv_plot_daily}
\title{Daily infections plot (RSV/VL simulator)}
\usage{
.rsv_plot_daily(households)
}
\arguments{
\item{households}{List of data frames from the RSV/VL simulator; each must
contain \code{role} and \code{infection_time}.}
}
\value{
A \code{ggplot} object.
}
\description{
Aggregates simulated household data to daily new infections by role.
}
\keyword{internal}
