% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-simulatehouseholds.R
\name{g_rescaled}
\alias{g_rescaled}
\title{Gaussian-like infectivity profile (normalized)}
\usage{
g_rescaled(t, peak_day, width)
}
\arguments{
\item{t}{Integer vector of times (e.g., days since infectious onset).}

\item{peak_day}{Integer; peak day.}

\item{width}{Numeric; spread around the peak.}
}
\value{
Numeric vector in [0,1] of length \code{t}.
}
\description{
Bell-shaped infectivity profile scaled to have maximum 1.
}
