% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{generate_contact_matrix_from_roles}
\alias{generate_contact_matrix_from_roles}
\title{Generate Household Contact Matrix from Role Rules}
\usage{
generate_contact_matrix_from_roles(current_roles, role_mixing_matrix)
}
\arguments{
\item{current_roles}{Vector of strings (e.g. c("adult", "adult", "infant"))}

\item{role_mixing_matrix}{4x4 Matrix with row/col names: infant, toddler, adult, elderly}
}
\value{
N x N matrix where N is length(current_roles)
}
\description{
Generate Household Contact Matrix from Role Rules
}
