% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_comput_fn.R
\name{get_theoretical_ct}
\alias{get_theoretical_ct}
\title{Compute Theoretical Ct Trajectory (Piecewise Linear)}
\usage{
get_theoretical_ct(t, p)
}
\description{
Compute Theoretical Ct Trajectory (Piecewise Linear)
}
\keyword{internal}
