% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_comput_fn.R
\name{get_theoretical_vl}
\alias{get_theoretical_vl}
\title{Compute Theoretical Log10 Viral Load Trajectory (Double Exponential)}
\usage{
get_theoretical_vl(t, p)
}
\description{
Compute Theoretical Log10 Viral Load Trajectory (Double Exponential)
}
\keyword{internal}
