% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HHBayes-package.R
\docType{package}
\name{HHBayes-package}
\alias{HHBayes}
\alias{HHBayes-package}
\title{HHBayes: Bayesian Household Transmission Modeling}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a streamlined pipeline to simulate household infection dynamics, estimate transmission parameters, and visualize epidemic timelines. Uses a Bayesian approach with Stan that models transmission probability as a Function of viral load, seasonality and infectivity, multiple infection episodes (reinfections), and waning immunity modeling.
}
\details{
\subsection{Funding}{

This work was supported by a grant from the National Institutes of Health
(R01AI137093). The content is solely the responsibility of the authors and
does not necessarily represent the official views of the National Institutes
of Health.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/keli5734/HHBayes}
  \item Report bugs at \url{https://github.com/keli5734/HHBayes/issues}
}

}
\author{
\strong{Maintainer}: Ke Li \email{ke.li.kl662@yale.edu}

Authors:
\itemize{
  \item Yiren Hou \email{yiren.hou@yale.edu}
}

}
\keyword{internal}
