% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{ode_system_func}
\alias{ode_system_func}
\title{ODE System for Within-Host Dynamics}
\usage{
ode_system_func(t, state, parms)
}
\arguments{
\item{t}{Numeric. Current time point.}

\item{state}{Numeric vector. Current state variables.}

\item{parms}{List. Model parameters for the ODE system.}
}
\value{
A list containing the derivatives of the state variables.
}
\description{
ODE System for Within-Host Dynamics
}
