% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_covariate_effects}
\alias{plot_covariate_effects}
\title{Plot Covariate Coefficients (Forest Plot)}
\usage{
plot_covariate_effects(fit, stan_data)
}
\arguments{
\item{fit}{A \code{stanfit} object returned by \code{fit_household_model}.}

\item{stan_data}{A list of data formatted by \code{prepare_stan_data}.}
}
\value{
A ggplot object displaying covariate effect estimates.
}
\description{
Plot Covariate Coefficients (Forest Plot)
}
