% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_posterior_distributions}
\alias{plot_posterior_distributions}
\title{Plot Posterior Distributions (Phi and Kappa)}
\usage{
plot_posterior_distributions(fit)
}
\arguments{
\item{fit}{A \code{stanfit} object returned by \code{fit_household_model}.}
}
\value{
A ggplot object displaying posterior violin plots.
}
\description{
Plot Posterior Distributions (Phi and Kappa)
}
