% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-simulatehouseholds.R
\name{simulate_viral_load_trajectory}
\alias{simulate_viral_load_trajectory}
\title{Simulate a log10 viral-load trajectory}
\usage{
simulate_viral_load_trajectory(t, v_p, t_p, lambda_g, lambda_d)
}
\arguments{
\item{t}{Integer vector; days relative to infection.}

\item{v_p}{Numeric; peak log10 VL.}

\item{t_p}{Integer; day of peak.}

\item{lambda_g, lambda_d}{Numeric; growth/decay rates.}
}
\value{
Numeric vector of log10 VL at times \code{t}.
}
\description{
Generates a unimodal log10 viral-load curve given peak level/day and
growth/decay rates.
}
