% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{summarize_attack_rates}
\alias{summarize_attack_rates}
\title{Summarize Attack Rates & Reinfections}
\usage{
summarize_attack_rates(sim_result)
}
\arguments{
\item{sim_result}{A list object returned by \code{simulate_multiple_households_comm}.}
}
\value{
A list containing four dataframes:
\describe{
\item{primary_overall}{Overall primary attack rate statistics.}
\item{primary_by_role}{Primary attack rate stratified by age group.}
\item{reinf_overall}{Overall summary of reinfection counts and rates.}
\item{reinf_by_role}{Reinfection counts stratified by age group.}
}
}
\description{
Calculates the Primary Attack Rate (proportion of people infected at least once)
and a separate summary of Reinfections (secondary episodes).
}
