% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenSyn.R
\name{print.GenSynResult}
\alias{print.GenSynResult}
\title{Print method for \code{GenSynResult}}
\usage{
\method{print}{GenSynResult}(x, ...)
}
\arguments{
\item{x}{A \code{GenSynResult} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{GenSynResult}
}
\examples{
\dontrun{
# Simulate and estimate
seasonal_forcing_list <- list(
  adult=rep(0.1,365), child=rep(0.1,365), elderly=rep(0.1,365), toddler=rep(0.1,365)
)
fit <- GenSyn(
  n_households=50,
  seasonal_forcing_list=seasonal_forcing_list, max_days=365,
  stan_chains=4, stan_iter=2000, stan_warmup=1000, stan_cores=4
)
print(fit)
}
}
